# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CalendarEventGameType = typeof CalendarEventGameType[keyof typeof CalendarEventGameType];


export const CalendarEventGameType = {
  campaign: 'campaign',
  module: 'module',
  oneshot: 'oneshot',
} as const;

export interface CalendarEvent {
  id: string;
  title: string;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  eventDate: string;
  startTime: string;
  gameType?: CalendarEventGameType;
  /** @nullable */
  catalogItemId?: string | null;
  status: string;
  /** @nullable */
  seats: number | null;
  description: string;
  system: string;
  format: string;
  location: string;
  duration: string;
  price: string;
  experience: string;
  age: string;
  playerPrep: string;
  archived: boolean;
  applicationUrl: string;
  recurrence: string;
  /**
     * @nullable
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
     */
  recurrenceUntil: string | null;
  excludedDates: string[];
}

export interface CalendarResponse {
  events: CalendarEvent[];
}

export interface GameSelection {
  id: string;
  title: string;
  date: string;
  dateLabel: string;
  time: string;
  status: string;
  price: string;
  place: string;
  location: string;
  preparation: string;
  revision: string;
}

export interface ApplicationSelectionResponse {
  game: GameSelection;
}

export interface ApplicationInput {
  submissionId: string;
  name: string;
  contact: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  players: number;
  format: string;
  place: string;
  experience?: string;
  system?: string;
  genres?: string;
  tone?: string;
  wishes?: string;
  schedule?: string;
  boundaries?: string;
  website?: string;
  consent: boolean;
  /** @nullable */
  eventId?: string | null;
  /** @nullable */
  occurrenceDate?: string | null;
  /** @nullable */
  eventRevision?: string | null;
  /** @maxLength 100 */
  utmSource?: string;
  /** @maxLength 100 */
  utmMedium?: string;
  /** @maxLength 100 */
  utmCampaign?: string;
  /** @maxLength 100 */
  utmContent?: string;
  /** @maxLength 100 */
  utmTerm?: string;
}

export interface ApplicationResponse {
  ok: boolean;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  notice?: string | null;
  /** @nullable */
  selectionChanged?: boolean | null;
  game?: GameSelection | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetApplicationSelectionParams = {
event: string;
/**
 * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
 */
date: string;
};