# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  submissionId: string;
  name: string;
  contact: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  players: number;
  format: string;
  place: string;
  experience?: string;
  system?: string;
  genres?: string;
  tone?: string;
  wishes?: string;
  schedule?: string;
  boundaries?: string;
  website?: string;
  consent: boolean;
  /** @nullable */
  eventId?: string | null;
  /** @nullable */
  occurrenceDate?: string | null;
  /** @nullable */
  eventRevision?: string | null;
  /** @maxLength 100 */
  utmSource?: string;
  /** @maxLength 100 */
  utmMedium?: string;
  /** @maxLength 100 */
  utmCampaign?: string;
  /** @maxLength 100 */
  utmContent?: string;
  /** @maxLength 100 */
  utmTerm?: string;
}