# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarEventGameType } from './calendarEventGameType';

export interface CalendarEvent {
  id: string;
  title: string;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  eventDate: string;
  startTime: string;
  gameType?: CalendarEventGameType;
  /** @nullable */
  catalogItemId?: string | null;
  status: string;
  /** @nullable */
  seats: number | null;
  description: string;
  system: string;
  format: string;
  location: string;
  duration: string;
  price: string;
  experience: string;
  age: string;
  playerPrep: string;
  archived: boolean;
  applicationUrl: string;
  recurrence: string;
  /**
     * @nullable
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
     */
  recurrenceUntil: string | null;
  excludedDates: string[];
}